refactor(agent): extract GitView file-row/diff components + helpers
GitView.tsx was 1098 lines. Pull the presentational pieces + pure helpers into
src/components/agent/git/ (no behaviour or styling change; className strings
preserved verbatim):

- git-helpers.ts — GitStatusData/GitFileEntry/GitLogData types, statusLabel/
  statusColor (pure), and the diffKey helper.
- FileSection.tsx / FileRow.tsx / InlineDiff.tsx — the changed-file section
  header, the per-file row (stage/unstage/discard + expandable diff), and the
  raw-diff renderer. (Inlinediff renamed to InlineDiff — it was file-local.)

Host GitView.tsx drops 1098 -> 914 (state, git-action handlers, and the render
composition remain). The action handlers were left inline — extracting them into
a hook is a larger, higher-risk change (heavy state threading) best done
separately. Only GitView is imported externally (AgentView), so call-site-safe.

Verified: type-check clean; eslint clean; full suite (1128) green.
